Parameter as double

+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/editor/parametervalues/PiParameterTableLabelProvider.java
@@ -128,7 +128,7 @@ public class PiParameterTableLabelProvider extends LabelProvider implements ITab
       case 5: // Expression Value Column
         if (parameter.isLocallyStatic()) {
           if (ExpressionEvaluator.canEvaluate(parameter, overrideExpression)) {
-            return Long.toString(
+            return Double.toString(
                 ExpressionEvaluator.evaluate(parameter, overrideExpression, this.scenario.getParameterValues().map()));
           }
           return overrideExpression;
